# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BUCT Media HUB",
    "description": "北京化工大学媒体资源管理系统后端接口。\n\n认证方式：大多数受保护接口使用 `Authorization: Bearer <token>`。\n前端常用登录接口为 `POST /api/v1/auth/login`，OpenAPI 工具链兼容登录接口为 `POST /api/v1/auth/token`。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/permissions/grant": {
      "post": {
        "tags": [
          "授权管理"
        ],
        "summary": "Grant Permission",
        "description": "授予用户权限\n\n管理员可以按学号给用户授权访问特定资源（如 Portrait 类别）。",
        "operationId": "grant_permission_api_v1_admin_permissions_grant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/permissions/user/{student_id}": {
      "get": {
        "tags": [
          "授权管理"
        ],
        "summary": "Get User Permissions",
        "description": "获取用户的权限列表\n\n按学号查询用户的所有授权记录。",
        "operationId": "get_user_permissions_api_v1_admin_permissions_user__student_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/{permission_id}": {
      "delete": {
        "tags": [
          "授权管理"
        ],
        "summary": "Revoke Permission",
        "description": "撤销权限",
        "operationId": "revoke_permission_api_v1_admin_permissions__permission_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/check/{student_id}": {
      "get": {
        "tags": [
          "授权管理"
        ],
        "summary": "Check User Permission",
        "description": "检查用户是否有特定权限",
        "operationId": "check_user_permission_api_v1_admin_permissions_check__student_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "category",
              "title": "Resource Type"
            }
          },
          {
            "name": "resource_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Portrait",
              "title": "Resource Key"
            }
          },
          {
            "name": "permission_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "view",
              "title": "Permission Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionGrantRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceTypeEnum",
            "default": "category"
          },
          "resource_key": {
            "type": "string",
            "title": "Resource Key",
            "default": "Portrait"
          },
          "permission_type": {
            "$ref": "#/components/schemas/PermissionTypeEnum",
            "default": "view"
          },
          "days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days",
            "default": 30
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "student_id"
        ],
        "title": "PermissionGrantRequest",
        "description": "授权请求 - 简化版\n\n管理员可以按学号授权访问特定类别。"
      },
      "PermissionList": {
        "properties": {
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionResponse"
            },
            "type": "array",
            "title": "Permissions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "permissions",
          "total"
        ],
        "title": "PermissionList",
        "description": "权限列表响应"
      },
      "PermissionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "user_student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Student Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_key": {
            "type": "string",
            "title": "Resource Key"
          },
          "permission_type": {
            "type": "string",
            "title": "Permission Type"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "resource_type",
          "resource_key",
          "permission_type",
          "start_time",
          "end_time",
          "is_active",
          "created_at",
          "note"
        ],
        "title": "PermissionResponse",
        "description": "权限响应"
      },
      "PermissionTypeEnum": {
        "type": "string",
        "enum": [
          "view",
          "download",
          "upload"
        ],
        "title": "PermissionTypeEnum",
        "description": "权限类型"
      },
...
      "ResourceTypeEnum": {
        "type": "string",
        "enum": [
          "category",
          "photo",
          "tag"
        ],
        "title": "ResourceTypeEnum",
        "description": "资源类型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "description": "使用 `POST /api/v1/auth/token` 获取 Bearer Token，或通过 `POST /api/v1/auth/login` 获取前端登录响应。"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发环境"
    },
    {
      "url": "/",
      "description": "相对路径，适合反向代理或 ApiFox 自定义环境"
    }
  ],
...
}